+++ apps/api/pi_dash/prompting/composer.py
@@ -476,6 +476,10 @@ def build_direct_turn(raw_prompt: str, run, issue=None) -> str:
             },
             "available_tools": run.tool_plan.get("tools", []),
             "unavailable_capabilities": run.tool_plan.get("unavailable_capabilities", []),
+            # True when the run may also carry toolsets whose tool names are not in
+            # ``tools`` (see ``pi_dash.ee.cloud_agent.toolsets``); the prompt has to
+            # say they exist because the capability list cannot name them.
+            "extra_toolsets": bool((getattr(run, "tool_plan", {}) or {}).get("extra_toolsets")),
             "limits": run.tool_plan.get("limits", {}),
         }
     composed = compose_cloud(recipes.KIND_DIRECT, workspace=workspace, project=project, context=context)
