+++ apps/api/pi_dash/prompting/context.py
@@ -395,6 +395,10 @@ def build_context(issue: Issue, run: AgentRun) -> Dict[str, Any]:
         },
         "available_tools": (getattr(run, "tool_plan", {}) or {}).get("tools", []),
         "unavailable_capabilities": (getattr(run, "tool_plan", {}) or {}).get("unavailable_capabilities", []),
+        # True when the run may also carry toolsets whose tool names are not in
+        # ``tools`` (see ``pi_dash.ee.cloud_agent.toolsets``); the prompt has to
+        # say they exist because the capability list cannot name them.
+        "extra_toolsets": bool((getattr(run, "tool_plan", {}) or {}).get("extra_toolsets")),
         "limits": (getattr(run, "tool_plan", {}) or {}).get("limits", {}),
         # Ticking schedule (None when the issue has no ticker row). Lets the
         # template explain the re-invocation cadence and remaining budget.
@@ -477,6 +481,10 @@ def build_scheduler_context(binding, run: AgentRun) -> Dict[str, Any]:
         },
         "available_tools": (getattr(run, "tool_plan", {}) or {}).get("tools", []),
         "unavailable_capabilities": (getattr(run, "tool_plan", {}) or {}).get("unavailable_capabilities", []),
+        # True when the run may also carry toolsets whose tool names are not in
+        # ``tools`` (see ``pi_dash.ee.cloud_agent.toolsets``); the prompt has to
+        # say they exist because the capability list cannot name them.
+        "extra_toolsets": bool((getattr(run, "tool_plan", {}) or {}).get("extra_toolsets")),
         "limits": (getattr(run, "tool_plan", {}) or {}).get("limits", {}),
         "scheduler_task_body": scheduler_task_body,
     }
